Fixes version serialization
